Feature/2

+++ src/main/java/com/team5/mapper/BoardMapper.java
@@ -3,5 +3,5 @@ package com.team5.mapper;
 import com.team5.domain.BoardVO;
 
 public interface BoardMapper {
-	public BoardVO read(int mi_num);
+	public BoardVO read(Long mi_num);
 }
